requirements.txt
ffmpeg-python==0.2.0
moviepy==1.0.3
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy<2 